--- a/vllm/envs.py
+++ b/vllm/envs.py
@@ -70,6 +70,7 @@ if TYPE_CHECKING:
     VLLM_DISABLED_KERNELS: List[str] = []
     VLLM_USE_V1: bool = False
     VLLM_ENABLE_V1_MULTIPROCESSING: bool = False
+    VLLM_STATS_ENGINE_POLLING_INTERVAL_S: int = 1
 
 
 def get_default_cache_root():
@@ -457,6 +458,10 @@ environment_variables: Dict[str, Callable[[], Any]] = {
     # If set, enable multiprocessing in LLM for the V1 code path.
     "VLLM_ENABLE_V1_MULTIPROCESSING":
     lambda: bool(int(os.getenv("VLLM_ENABLE_V1_MULTIPROCESSING", "0"))),
+
+    # Interval in seconds to poll the engine stats.
+    "VLLM_STATS_ENGINE_POLLING_INTERVAL_S":
+    lambda: int(os.getenv("VLLM_STATS_ENGINE_POLLING_INTERVAL_S", "1")),
 }
 
 # end-env-vars-definition