--- a/vllm/v1/engine/__init__.py
+++ b/vllm/v1/engine/__init__.py
@@ -7,7 +7,6 @@ import msgspec
 from vllm.lora.request import LoRARequest
 from vllm.multimodal import MultiModalDataDict, MultiModalPlaceholderDict
 from vllm.sampling_params import RequestOutputKind, SamplingParams
-from vllm.v1.engine.stats import EngineCoreStats
 
 
 @dataclass
@@ -68,12 +67,15 @@ class EngineCoreOutputs(msgspec.Struct,
     # [num_reqs]
     outputs: List[EngineCoreOutput]
 
-    stats: Optional[EngineCoreStats] = None
-
 
 @dataclass
 class EngineCoreProfile:
     is_start: bool
+
+
+@dataclass
+class EngineCoreStatsRequest:
+    pass
 
 
 class EngineCoreRequestType(enum.Enum):
@@ -84,3 +86,4 @@ class EngineCoreRequestType(enum.Enum):
     ADD = b'\x00'
     ABORT = b'\x01'
     PROFILE = b'\x02'
+    STATS = b"\x03"