## Requirements and Goals
1. Stats collection should work with both the legacy and new multiprocess architecture.
2. Stats collection should be possible when engine core is busy in doing GPU work (or stuck at something) so we get observability metrics even when engine core is stuck.
3. Stats collection should take into account the entire lifecycle of a request, i.e. the input processing time.
5. Stats collection should incur minimal overheads on runtime/memory --> should likely be decoupled from the engine core loop. 

## Proposal and Design
1. Introducing `RequestStats` and `RequestStatsUpdate` to capture requests updates for metrics. It captures the entire request lifecycle (arrival -> input process -> engine core prefill/decode -> [preempt] -> detokenized). 
2. Introducing a `EngineStatsAgent` that lives with the `EngineCore` to collect stats updates from various core components and scheduler/model outputs
6. Introducing a `EngineStatsManager` class that lives on the `AsyncLLM` engine that aggregates stats from the agent.
7. The manager and agent communicates through the ZMQ sockets for async stats collection. 

```
Main Process                                  Engine Core Process
+----------------------+                      +----------------------+
|      AsyncLLM        |                     |    EngineCoreProc    |
|  +-----------------+ |                     |  +-----------------+  |
|  |EngineStatsManager| |                    |  |EngineStatsAgent |  |
|  +-----------------+ |                     |  +-----------------+  |
|                      |                     |                      |
|  [AsyncIO Event Loop]|     ZMQ Sockets     |  [Background Threads]|
|  +-----------------+ |                     |  +-----------------+  |
|  | Output Handler  |<|---------------------|->| Output Socket   |  |
|  +-----------------+ |                     |  +-----------------+  |
|                      |                     |                      |
|  +-----------------+ |                     |  +-----------------+  |
|  | Stats Handler   |<|---------------------|->| Stats Socket    |  |
|  +-----------------+ |                     |  +-----------------+  |
|                      |                     |                      |
|                      |                     |  +-----------------+  |
|                      |---------------------|->| Input Socket    |  |
|                      |                     |  +-----------------+  |
|                      |                     |                      |
|                      |                     |  [Main Thread]       |
|                      |                     |  +-----------------+  |
|                      |                     |  |   Busy Loop    |  |
|                      |                     |  +-----------------+  |
+----------------------+                      +----------------------+
```

- **On Backend (EngineCore Process)**:
  - New `EngineStatsAgent` collects metrics in a thread-safe manner
  - Stats are sent to frontend via ZMQ sockets
  - Runs in separate background threads to minimize impact on model execution

- **On Frontend (Main Process)**:
  - `EngineStatsManager` aggregates stats from backend as well as collect stats on the engine frontend.
  - AsyncIO-based stats polling with configurable interval
  - Compatible with existing V0 StatsLogger interface 





## Benchmark Results
TLDR: There's no observable perf regression when stats collection happens every 1 sec for the below workloads comparing to stats collection turned off. 


### Throughput 
- PR with stats logging every 1 sec: 13.7 requests/s
- PR without: 13.7 requests/s
```
VLLM_USE_V1=1 python benchmarks/benchmark_throughput.py --input-len 1024 --output-len 512 --num-prompts 1000 --async-engine \
--disable-log-requests [--disable-log-stat]
```


### Latency

```
============ Serving Benchmark Result ============
Metric                                  | Main       | PR with stats | PR without stats
----------------------------------------------------------------------------------------
Successful requests                     | 998        | 998           | 998
Benchmark duration (s)                  | 32.93      | 32.85         | 32.21
Total input tokens                      | 223244     | 223244        | 223244
Total generated tokens                  | 209009     | 209177        | 209029
Request throughput (req/s)              | 30.31      | 30.38         | 30.98
Output token throughput (tok/s)         | 6347.79    | 6366.81       | 6488.77
Total Token throughput (tok/s)          | 13127.92   | 13161.77      | 13418.80
---------------Time to First Token----------------
Mean TTFT (ms)                          | 3178.80    | 3134.97       | 3171.81
Median TTFT (ms)                        | 2980.97    | 3053.03       | 2960.46
P99 TTFT (ms)                           | 5466.19    | 5611.05       | 5377.22
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms)                          | 172.70     | 180.67        | 160.93
Median TPOT (ms)                        | 124.75     | 127.83        | 120.60
P99 TPOT (ms)                           | 438.38     | 599.12        | 409.44
---------------Inter-token Latency----------------
Mean ITL (ms)                           | 75.22      | 77.18         | 72.43
Median ITL (ms)                         | 36.60      | 36.04         | 36.26
P99 ITL (ms)                            | 852.84     | 908.12        | 806.74
==================================================

```