@rickyyx 2024-11-27
So in order to support the below:
1. Stats collection should be possible when engine core is busy (or stuck)
2. Stats collection should incur minimal overheads --> happens on the frontend of the engine. 

I am proposing the below changes:
1. Introducing a `EngineStatsAgent` that lives with the `EngineCore` to collect stats from various core components and scheduler/model outputs
3. Introducing a `EngineStatsManager` class that lives on the `AsyncLLM` engine that aggregates stats from the agent.
4. The manager and agent communicates through the ZMQ sockets for async stats collection. 

```
Main Process                                  Engine Core Process
+----------------------+                      +----------------------+
|      AsyncLLM        |                     |    EngineCoreProc    |
|  +-----------------+ |                     |  +-----------------+  |
|  |EngineStatsManager| |                    |  |EngineStatsAgent |  |
|  +-----------------+ |                     |  +-----------------+  |
|                      |                     |                      |
|  [AsyncIO Event Loop]|     ZMQ Sockets     |  [Background Threads]|
|  +-----------------+ |                     |  +-----------------+  |
|  | Output Handler  |<|---------------------|->| Output Socket   |  |
|  +-----------------+ |                     |  +-----------------+  |
|                      |                     |                      |
|  +-----------------+ |                     |  +-----------------+  |
|  | Stats Handler   |<|---------------------|->| Stats Socket    |  |
|  +-----------------+ |                     |  +-----------------+  |
|                      |                     |                      |
|                      |                     |  +-----------------+  |
|                      |---------------------|->| Input Socket    |  |
|                      |                     |  +-----------------+  |
|                      |                     |                      |
|                      |                     |  [Main Thread]       |
|                      |                     |  +-----------------+  |
|                      |                     |  |   Busy Loop    |  |
|                      |                     |  +-----------------+  |
+----------------------+                      +----------------------+
```